requirements.txt
# 项目管理 Agent MVP 依赖包
# 后端核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 定时任务
apscheduler==3.10.4

# 工具库
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
isort==5.12.0
ruff==0.1.6
